Split `CourseEditorScreen` to sub-components
